Homework 5-31-2017 Else Clauses
Here is my homework for the else clauses. I will be updating this read me file as I go through the rest of this class. It is replacing my "explanation" script in the SummerDGM1600 project. But I will leave that there for now because it has my earlier assignments in it.